+++ history.md
@@ -1,6 +1,9 @@
+## 0.3.0
+- Implement Do#complete
+- Make error/success callbacks optional if complete is used
+
 ## 0.2.0
 - Fix sync callbacks issue, more info in Do#done api description.
 
 ## 0.1.0
-
 - Ensured success callback can be triggered only once.